refactor #29: dto 패키지 명확하게 구분
